requirements.txt
langchain==0.0.228
pinecone_client==2.2.2
python-dotenv==1.0.0
streamlit==1.24.0
